Fixed a bug with which heuristic is used. Finished documenting.
Stupidly forgot to do a switch in the heuristics choice in the search. The
program was initializing the first state with the correct heuristic but each
subsequent heuristic was the Manhattan heuristic. Needless to say the results
were skewed because of this. I should have fully documented the code I added.