Changed A* to be usable for both searches.
The A* and Greedy Best First searches are virtually identical algorithms with
the only difference being their comparator evaluation. By merely adding an if-
else I can use the astar method to run both searches as well as both heuristics
that are required for this assignment. Yay for less lines of code. Yay for
finishing documentation. Yay for this assignment is done.