Fixed the coding style to match my preferences.
I need to keep in mind that if there are errors I removed the {} in for-loops.
The for-loops do have nested ifs that are only "one level in" but have line
breaks to make the code easier to read in text editors. If this is an error,
I'll readd the {} to the code.